+++ @xen-orchestra/lite/CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - [VM/New] Added secureBoot support (PR [#9423](https://github.com/vatesfr/xen-orchestra/pull/9423))
 - [Dashboard] Fix reactivity of dashboard (PR [#9378](https://github.com/vatesfr/xen-orchestra/pull/9378))
+- [Header] Fix menu button on large screen (PR [#9558](https://github.com/vatesfr/xen-orchestra/pull/9558))
 
 ## **0.19.0** (2026-02-26)